requirements.txt
# Streamlit Framework
streamlit>=1.31.0

# Auth0 Authentication
authlib>=1.3.0
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# Database (SQLite for now, can upgrade to PostgreSQL later)
sqlalchemy>=2.0.25

# Data Processing
pandas>=2.2.0

# Additional utilities
cryptography>=42.0.0

# Flask API for Arduino data ingestion
flask>=3.0.0
flask-cors>=4.0.0
